[amebadplus][zephyr] fix bugs inside i2s s3
* Modify: fix bugs to pass test-speed

ChangeLogs:
1) Replace fifo-empty-timer by fifo-empty-irq. The irq is former used by
dsp-mode, therefore it has not been used before. The function of this
irq works well in dma mode. So use this irq to judge the end of TX SPORT.
2) test-api: error test_i2s_config_attempt_in_wrong_state stops i2s and
configure the next testcase immediately, which is supposed to wait the
STOPPING status. So reserved the wait in configure func.
3) In single irq loopback mode, the i2s-0 cannot be set as slave. Judge
this state and set in configure. In test-api, upper layer set rx first
then tx, so this mode is master, but in test-speed upper layer config tx
first then rx, so this mode changed to slave, which is not true.
4) Delete I2S_INSIDE_LOOPBACK which is only used in debug.
5) Delete src_addr_increment and dst_addr_increment in initialize,
because these two parameters are not used and do not need to init at
all.

Change-Id: Iecee7b9dc2fffc51ef07205b9b579cfbb80fb7da